[toolchain] `install_extra_deps.py` further fixes
This PR does some additional fixes to `install_extra_deps.py`, making it
more stable for use. The main things fixed in this change are:

 * `DEPS` is now parsed as read-only, to avoid any unwanted
   side-effects.
 * `.vpython3` is being introduced `install_extra_deps.py`, as this
   script is expected to be run in place.
 * The script is being renamed to `install_extra_deps.py`, as this is
   more self-descriptive.

The values committed now in the `EXTRA_DEPS` set are valid data.

Bug: https://github.com/brave/brave-browser/issues/55812
