+++ tools/cr/toolchains/.vpython3
@@ -1,0 +1,15 @@
+# Copyright (c) 2026 The Brave Authors. All rights reserved.
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this file,
+# You can obtain one at https://mozilla.org/MPL/2.0/.
+
+python_version: "3.11"
+
+wheel: <
+  name: "infra/python/wheels/httplib2-py3"
+  version: "version:0.22.0"
+>
+wheel: <
+  name: "infra/python/wheels/pyparsing-py3"
+  version: "version:3.2.0"
+>
